test(filters): consolidate null ordering coverage
Keep NULL placement compilation, eager-loading pagination, and compiled-cache regressions beside the existing OrderBy integration tests instead of in a separate unit-test module. Retain independent fixtures and unit markers so dialect compilation coverage does not require a running database.

All 98 selected filter tests and the full 292-file mypy check pass. This only reorganizes tests and leaves runtime behavior unchanged.
